Rename `WithdrawalTestContext`.
